# doc/openapi.yaml
openapi: 3.0.0
info:
  title: Playzone API
  version: '0.1'
servers:
  - url: 'http://localhost:8000'
paths:
  /api/auth/sign-in:
    parameters: []
    post:
      summary: Se connecter
      operationId: post-api-auth-sign-in
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                required:
                  - token
              examples:
                Example 1:
                  value:
                    token: BEARER ...
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorForm'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorForbidden'
      description: ''
      tags:
        - auth
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                login:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
              required:
                - login
                - password
  /api/auth/sign-out:
    post:
      summary: Se déconnecter
      operationId: post-api-auth-sign-out
      responses:
        '204':
          description: No Content
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorUnauthorized'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorForbidden'
      tags:
        - auth
      security:
        - ADMIN: []
        - USER: []
  /api/auth/change-password:
    post:
      summary: Effectuer une demande de changement de mot de passe
      operationId: post-api-auth-recover-password
      responses:
        '204':
          description: No Content
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorForm'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
              required:
                - email
            examples: {}
        description: ''
      tags:
        - auth
      description: |-
        Envoie un email de changement de mot de passe à l'utilisateur.
        Ce mail contient un lien vers une page pour changer son mot de passe.
    parameters: []
  '/api/auth/change-password/{token}':
    post:
      summary: Modifier le mot de passe
      operationId: post-api-auth-change-password
      responses:
        '204':
          description: No Content
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorForm'
        '404':
          description: Not Found
          content: {}
      description: ''
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                password:
                  type: string
                  format: password
              required:
                - password
      tags:
        - auth
    parameters:
      - schema:
          type: string
          example: 01BX5ZZKBKACTAV9WEVGEMMVRY
        name: token
        in: path
        required: true
        description: token identifiant la demande de changement de mot de passe
  /api/auth/profile:
    get:
      summary: Récupérer l'utilisateur connecté
      tags:
        - auth
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorUnauthorized'
      operationId: get-api-auth-profile
      security:
        - ADMIN: []
        - USER: []
...
components:
  schemas:
    Profile:
      title: Profile
      type: object
      x-examples: {}
      properties:
        id:
          type: string
          example: 01BX5ZZKBKACTAV9WEVGEMMVRY
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
        email:
          type: string
          format: email
      required:
        - id
        - firstName
        - lastName
        - email
...
    ErrorForm:
      title: ErrorForm
      type: object
      properties:
        status:
          type: integer
          example: 400
        path:
          type: string
          example: /api/endpoint
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: name
              message:
                type: string
                example: required
            required:
              - field
              - message
      required:
        - status
        - path
        - errors
...
    ErrorUnauthorized:
      title: ErrorUnauthorized
      type: object
      properties:
        status:
          type: integer
          example: 401
        path:
          type: string
          example: /api/endpoint
        error:
          type: string
          example: Unauthorized
      required:
        - status
        - path
        - error
    ErrorForbidden:
      title: ErrorForbidden
      type: object
      properties:
        status:
          type: integer
          example: 403
        path:
          type: string
          example: /api/endpoint
        error:
          type: string
          example: Forbidden
      required:
        - status
        - path
        - error
...
  securitySchemes:
    ADMIN:
      type: http
      scheme: bearer
      description: ''
    USER:
      type: http
      scheme: bearer
...